Refactor: contributors로 링크 및 이름 수정 & List는 따로 분리
