+++ app/contributors/page.tsx
@@ -1,5 +1,31 @@
-const page = () => {
-  return <div>contributors page</div>;
+'use client';
+import React, { useEffect, useState } from 'react';
+import Title from '@/components/contributors/Title';
+import ContributorCardList from '@/components/contributors/ContributorCardList';
+
+const Page = () => {
+  const [isScrolled, setIsScrolled] = useState(false);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setIsScrolled(true);
+    };
+
+    document.addEventListener('scroll', handleScroll);
+    return () => {
+      document.removeEventListener('scroll', handleScroll);
+    };
+  }, []);
+
+  return (
+    <div className="relative flex flex-col mb-[-160px] dt:mb-[-194px]">
+      {isScrolled && (
+        <div className="flex w-full h-[64px] fixed z-10 blur-lg backdrop-blur-sm top-0" />
+      )}
+      <Title />
+      <ContributorCardList />
+    </div>
+  );
 };
 
-export default page;
+export default Page;
