+++ components/contributors/ContributorList.tsx
@@ -1,0 +1,72 @@
+import JY from '@/public/image/contributors/JY.avif';
+import GR from '@/public/image/contributors/GR.avif';
+import SW from '@/public/image/contributors/SW.avif';
+import GY from '@/public/image/contributors/GY.avif';
+import SY from '@/public/image/contributors/SY.avif';
+import JE from '@/public/image/contributors/JE.avif';
+import GH from '@/public/image/contributors/GH.avif';
+
+export const ContributorList = [
+  {
+    image: GY,
+    name: '서가영',
+    role: '프론트엔드',
+    year: '20',
+    school: '홍익대학교 컴퓨터공학과 21학번',
+    githubUrl: 'https://github.com/caminobelllo',
+    githubName: '@caminobelllo',
+  },
+  {
+    image: SY,
+    name: '심수연',
+    role: '프론트엔드 · 백엔드',
+    year: '19',
+    school: '홍익대학교 컴퓨터공학과 20학번',
+    githubUrl: 'https://github.com/letthem',
+    githubName: '@letthem',
+  },
+  {
+    image: JE,
+    name: '염지은',
+    role: '프론트엔드 · 백엔드',
+    year: '20',
+    school: '홍익대학교 컴퓨터공학과 21학번',
+    githubUrl: 'https://github.com/yumzen',
+    githubName: '@yumzen',
+  },
+  {
+    image: GH,
+    name: '임가현',
+    role: '프론트엔드',
+    year: '20',
+    school: '홍익대학교 컴퓨터공학과 21학번',
+    githubUrl: 'https://github.com/limgahyun',
+    githubName: '@limgahyun',
+  },
+  {
+    image: JY,
+    name: '강지윤',
+    role: '백엔드',
+    year: '20',
+    school: '홍익대학교 컴퓨터공학과 21학번',
+    githubUrl: 'https://github.com/kjiyun',
+    githubName: '@kjiyun',
+  },
+  {
+    image: SW,
+    name: '박상욱',
+    role: '백엔드',
+    year: '18',
+    school: '홍익대학교 컴퓨터공학과 19학번',
+    githubUrl: 'https://github.com/woogieon8on',
+    githubName: '@woogieon8on',
+  },
+  {
+    image: GR,
+    name: '구름',
+    role: '디자인',
+    school: '서울여자대학교 디지털미디어학과 21학번',
+    githubUrl: '',
+    githubName: 'rnfma5856@gmail.com',
+  },
+];
